+++ Renderer/Renderer/SceneNodes/ModelSceneNode.cs
@@ -142,7 +142,7 @@ namespace ValveResourceFormat.Renderer.SceneNodes
             LoadAnimations(model, embeddedAnimationsOnly: isWorldPreview);
 
             SetCharacterEyeRenderParams();
-            TwistConstraints = ParseTwistConstraints(model);
+            AnimationController.TwistConstraints = ParseTwistConstraints(model);
         }
 
         readonly struct CharacterEyeParameters
@@ -236,7 +236,7 @@ namespace ValveResourceFormat.Renderer.SceneNodes
                 return;
             }
 
-            ApplyConstraints(context);
+            AnimationController.ApplyConstraints();
 
             if (IsAnimated)
             {
@@ -731,299 +731,6 @@ namespace ValveResourceFormat.Renderer.SceneNodes
 
             return [.. constraints];
         }
-
-        /// <summary>
-        /// Applies tilt-twist constraints to the animated skeleton. Called after animation controller updates.
-        /// Override this method to implement constraint logic or customize debug visualization.
-        /// </summary>
-        protected virtual void ApplyConstraints(Scene.UpdateContext context)
-        {
-            if (TwistConstraints.Length == 0)
-            {
-                return;
-            }
-
-            var skeleton = AnimationController.FrameCache.Skeleton;
-            var yOffset = 0.05f;
-
-            foreach (var constraint in TwistConstraints)
-            {
-                // Evaluate constraint and display debug info
-                var evalResult = EvaluateTiltTwistConstraint(constraint, skeleton, AnimationController.Pose);
-
-                context.TextRenderer.AddTextRelative($"Constraint: {constraint.Name}", 0.02f, yOffset, 11f, Color32.White, context.Camera);
-                yOffset += 0.02f;
-
-                // Display target info
-                foreach (var target in constraint.Targets)
-                {
-                    var boneName = StringToken.GetKnownString(target.BoneHash);
-                    var bone = skeleton[target.BoneHash];
-
-                    if (bone != null)
-                    {
-                        context.TextRenderer.AddTextRelative($"  Target: {boneName} (w={target.Weight:F2})", 0.02f, yOffset, 11f, new Color32(0f, 1f, 1f, 1f), context.Camera);
-                    }
-                    else
-                    {
-                        context.TextRenderer.AddTextRelative($"  Target: {boneName} NOT FOUND", 0.02f, yOffset, 11f, Color32.Red, context.Camera);
-                    }
-                    yOffset += 0.02f;
-                }
-
-                // Display twist angle
-                context.TextRenderer.AddTextRelative($"  Twist: {evalResult.TwistAngleDegrees:F1}deg", 0.02f, yOffset, 11f, Color32.Yellow, context.Camera);
-                yOffset += 0.02f;
-
-                // Display slave info
-                foreach (var slave in constraint.Slaves)
-                {
-                    var boneName = StringToken.GetKnownString(slave.BoneHash);
-                    var bone = skeleton[slave.BoneHash];
-
-                    if (bone != null)
-                    {
-                        context.TextRenderer.AddTextRelative($"  Slave: {boneName} (w={slave.Weight:F2})", 0.02f, yOffset, 11f, Color32.Green, context.Camera);
-                    }
-                    else
-                    {
-                        context.TextRenderer.AddTextRelative($"  Slave: {boneName} NOT FOUND", 0.02f, yOffset, 11f, Color32.Red, context.Camera);
-                    }
-                    yOffset += 0.02f;
-                }
-
-                yOffset += 0.01f; // Space between constraints
-            }
-        }
-
-        private readonly struct ConstraintEvaluation
-        {
-            public Quaternion TwistRotation { get; init; }
-            public float TwistAngleDegrees { get; init; }
-        }
-
-        /// <summary>
-        /// Evaluates a tilt-twist constraint by computing the twist rotation from target bones.
-        /// Applies the twist to slave bones and updates the pose hierarchy.
-        /// </summary>
-        private static ConstraintEvaluation EvaluateTiltTwistConstraint(TiltTwistConstraint constraint, Skeleton skeleton, Span<Matrix4x4> pose)
-        {
-            var twistAxis = GetAxisVector(constraint.TargetAxis);
-            var accumulatedTwist = Quaternion.Identity;
-            var totalWeight = 0f;
-
-            // Get weighted average twist from all targets
-            foreach (var target in constraint.Targets)
-            {
-                var bone = skeleton[target.BoneHash];
-                if (bone == null || target.Weight <= 0f)
-                {
-                    continue;
-                }
-
-                // Get local rotation relative to parent
-                var localRotation = GetLocalRotation(bone, pose);
-
-                // Remove the bind pose to get only the animated delta
-                var bindRotation = bone.Angle;
-                var deltaRotation = Quaternion.Multiply(Quaternion.Inverse(bindRotation), localRotation);
-
-                // Apply target offset
-                deltaRotation = Quaternion.Multiply(target.Offset, deltaRotation);
-
-                // Decompose rotation into twist (around axis) and swing (perpendicular)
-                var twist = DecomposeTwistRotation(deltaRotation, twistAxis);
-
-                // Accumulate twist using weighted slerp (proper quaternion averaging)
-                if (totalWeight == 0f)
-                {
-                    accumulatedTwist = Quaternion.Slerp(Quaternion.Identity, twist, target.Weight);
-                }
-                else
-                {
-                    var newWeight = target.Weight / (totalWeight + target.Weight);
-                    accumulatedTwist = Quaternion.Slerp(accumulatedTwist, twist, newWeight);
-                }
-
-                totalWeight += target.Weight;
-            }
-
-            // Calculate twist angle for display
-            var twistAngle = 2f * MathF.Acos(Math.Clamp(MathF.Abs(accumulatedTwist.W), 0f, 1f));
-            var twistAngleDegrees = twistAngle * 180f / MathF.PI;
-
-            // Apply twist to slave bones
-            foreach (var slave in constraint.Slaves)
-            {
-                var bone = skeleton[slave.BoneHash];
-                if (bone == null || slave.Weight <= 0f)
-                {
-                    continue;
-                }
-
-                // Get local rotation
-                var localRotation = GetLocalRotation(bone, pose);
-
-                // Apply weighted twist to the local rotation directly
-                var weightedTwist = Quaternion.Slerp(Quaternion.Identity, accumulatedTwist, slave.Weight);
-                var newLocalRotation = Quaternion.Multiply(weightedTwist, localRotation);
-
-                // Apply the new local rotation and rebuild hierarchy
-                ApplyLocalRotationAndRebuildChildren(bone, newLocalRotation, pose);
-            }
-
-            return new ConstraintEvaluation
-            {
-                TwistRotation = accumulatedTwist,
-                TwistAngleDegrees = twistAngleDegrees
-            };
-        }
-
-        /// <summary>
-        /// Extracts the local rotation of a bone relative to its parent from world-space pose.
-        /// </summary>
-        private static Quaternion GetLocalRotation(Bone bone, Span<Matrix4x4> pose)
-        {
-            Matrix4x4.Decompose(pose[bone.Index], out _, out var worldRotation, out _);
-
-            if (bone.Parent != null)
-            {
-                Matrix4x4.Decompose(pose[bone.Parent.Index], out _, out var parentRotation, out _);
-                return Quaternion.Multiply(Quaternion.Inverse(parentRotation), worldRotation);
-            }
-
-            return worldRotation;
-        }
-
-        /// <summary>
-        /// Applies a new local rotation to a bone and recursively updates world transforms for all children.
-        /// </summary>
-        private static void ApplyLocalRotationAndRebuildChildren(Bone bone, Quaternion newLocalRotation, Span<Matrix4x4> pose)
-        {
-            // Get parent world transform
-            var parentTransform = bone.Parent != null ? pose[bone.Parent.Index] : Matrix4x4.Identity;
-
-            // Extract current local translation and scale from world transform
-            if (bone.Parent != null)
-            {
-                // Transform world position to parent local space to get local translation
-                Matrix4x4.Decompose(pose[bone.Index], out var worldScale, out _, out var worldTranslation);
-                Matrix4x4.Decompose(parentTransform, out _, out _, out var parentTranslation);
-                
-                var localTranslation = Vector3.Transform(worldTranslation - parentTranslation, Matrix4x4.CreateFromQuaternion(Quaternion.Inverse(GetRotation(parentTransform))));
-
-                // Build new local transform
-                var localTransform = Matrix4x4.CreateScale(worldScale)
-                    * Matrix4x4.CreateFromQuaternion(newLocalRotation)
-                    * Matrix4x4.CreateTranslation(localTranslation);
-
-                // Compute new world transform
-                pose[bone.Index] = localTransform * parentTransform;
-            }
-            else
-            {
-                // Root bone - just update rotation in world space
-                Matrix4x4.Decompose(pose[bone.Index], out var scale, out _, out var translation);
-                pose[bone.Index] = Matrix4x4.CreateScale(scale)
-                    * Matrix4x4.CreateFromQuaternion(newLocalRotation)
-                    * Matrix4x4.CreateTranslation(translation);
-            }
-
-            // Recursively rebuild all children
-            foreach (var child in bone.Children)
-            {
-                RebuildChildWorldTransform(child, pose);
-            }
-        }
-
-        /// <summary>
-        /// Recursively rebuilds world transforms for a bone and its children based on parent's updated transform.
-        /// </summary>
-        private static void RebuildChildWorldTransform(Bone bone, Span<Matrix4x4> pose)
-        {
-            var parentTransform = bone.Parent != null ? pose[bone.Parent.Index] : Matrix4x4.Identity;
-
-            // Get the bone's current local rotation (preserve animation)
-            var localRotation = GetLocalRotation(bone, pose);
-
-            // Extract local translation from world space
-            Matrix4x4.Decompose(pose[bone.Index], out var worldScale, out _, out var worldTranslation);
-            
-            Vector3 localTranslation;
-            if (bone.Parent != null)
-            {
-                Matrix4x4.Decompose(parentTransform, out _, out var parentRotation, out var parentTranslation);
-                var relativeTranslation = worldTranslation - parentTranslation;
-                localTranslation = Vector3.Transform(relativeTranslation, Matrix4x4.CreateFromQuaternion(Quaternion.Inverse(parentRotation)));
-            }
-            else
-            {
-                localTranslation = worldTranslation;
-            }
-
-            // Rebuild local transform with preserved animation
-            var localTransform = Matrix4x4.CreateScale(worldScale)
-                * Matrix4x4.CreateFromQuaternion(localRotation)
-                * Matrix4x4.CreateTranslation(localTranslation);
-
-            // Compute new world transform
-            pose[bone.Index] = localTransform * parentTransform;
-
-            // Recursively rebuild children
-            foreach (var child in bone.Children)
-            {
-                RebuildChildWorldTransform(child, pose);
-            }
-        }
-
-        /// <summary>
-        /// Extracts rotation from a transformation matrix.
-        /// </summary>
-        private static Quaternion GetRotation(Matrix4x4 matrix)
-        {
-            Matrix4x4.Decompose(matrix, out _, out var rotation, out _);
-            return rotation;
-        }
-
-        /// <summary>
-        /// Decomposes a quaternion into its twist component around the given axis.
-        /// </summary>
-        private static Quaternion DecomposeTwistRotation(Quaternion rotation, Vector3 axis)
-        {
-            // Project rotation axis onto constraint axis
-            var rotationAxis = new Vector3(rotation.X, rotation.Y, rotation.Z);
-
-            if (rotationAxis.LengthSquared() < 1e-6f)
-            {
-                return Quaternion.Identity;
-            }
-
-            var projection = Vector3.Dot(rotationAxis, axis) * axis;
-
-            // Reconstruct twist quaternion
-            var twist = new Quaternion(projection.X, projection.Y, projection.Z, rotation.W);
-            var length = twist.Length();
-
-            if (length < 1e-6f)
-            {
-                return Quaternion.Identity;
-            }
-
-            return Quaternion.Normalize(twist);
-        }
-
-        /// <summary>
-        /// Converts axis index (0=X, 1=Y, 2=Z) to unit vector.
-        /// </summary>
-        private static Vector3 GetAxisVector(int axisIndex)
-        {
-            return axisIndex switch
-            {
-                0 => Vector3.UnitX,
-                1 => Vector3.UnitY,
-                2 => Vector3.UnitZ,
-                _ => Vector3.UnitX,
-            };
-        }
     }
 }
+
